backend/requirements.txt
tensorflow==2.15.0
numpy==1.26.2
pandas==2.1.3
pillow==10.1.0
scikit-learn==1.3.2
matplotlib==3.8.2
flask==3.0.0
flask-cors==4.0.0
opencv-python==4.8.1.78
seaborn==0.13.0
tqdm==4.66.1
